fixed negative results bug and switched to using longs where an integer value could get above 100,000

+++ src/main/java/com/behaviorlogger/controllers/NotesController.java
@@ -121,7 +121,7 @@ public class NotesController
     // Ctrl/cmd+T inserts the current session timestamp
     if (evt.isShortcutDown() && code == KeyCode.T) {
       int index = textArea.getCaretPosition();
-      int count = manager.count();
+      long count = manager.count();
       String timestamp = BehaviorLoggerUtil.millisToTimestamp( count );
       String msg = String.format( "[%s] ", timestamp );
       textArea.insertText( index, msg );
